client/requirements.txt
# GUI Dependencies
Tkinter # Usually pre-installed with Python

# Communication Dependencies
pyserial
python-mbedtls